requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.8.0
djangorestframework==3.12.4
djongo==1.3.7
pillow==11.0.0
pymongo==3.11.4
pytz==2024.2
setuptools==75.5.0
sqlparse==0.2.4
